requirements.txt
Django==2.2.5
django-admin-row-actions==0.0.5
django-colorfield
python-dotenv==0.10.1
psycopg2-binary
pytest
pytest-django
pytest-cov
factory_boy
faker
Pillow